// frontend/src/pages/shared/interface/user.interface.ts
export interface IUser {
  bio?: string;
  birthday?: Date;
  coverPostId?: number;
  createdAt: Date;
  email: string;
  location?: string;
  name: string;
  password: string;
  position?: string;
  profilePostId?: number;
  school?: string;
  workplace?: string;
  coverImagePostId?: number;
  profileImageUrl?: string;
}